Footprint can be pointer to specific location, demo ready.
Note:
The compass package also depended on rxdart package. In this case so The corresponding version was downgraded to 0.23.1, which leading a new issues , I have to run the application with "--no-sound-null-safety" other it will print error message "Non-safety issues".
This version is demo ready on the fontys TQ location.